+++ tools/accuracy_checker/accuracy_checker/data_readers/data_reader.py
@@ -289,9 +289,6 @@ class NiftiImageReader(BaseReader):
             image = np.expand_dims(image, -1)
         image = np.transpose(image, (3, 0, 1, 2))
 
-        order = [1, 3, 0, 2] if image.shape[0] == 4 else [0]
-        image = image[order, :, :, :]
-
         return image
