+++ tools/accuracy_checker/accuracy_checker/preprocessor/geometric_transformations.py
@@ -650,63 +650,3 @@ class ImagePyramid(Preprocessor):
         image.metadata.update({'multi_infer': True, 'scales': scales})
 
         return image
-
-
-class Resize3D(Preprocessor):
-    __provider__ = 'resize3d'
-
-    @classmethod
-    def parameters(cls):
-        parameters = super().parameters()
-        parameters.update({
-            'size': BaseField(optional=True, description='Specifies resize'),
-        })
-        return parameters
-
-    def configure(self):
-        self.shape = get_or_parse_value(self.config.get('size'), (128, 128, 128), casting_type=int)
-        self.shape = (4,) + self.shape
-
-    def process(self, image, annotation_meta=None):
-        data = np.asarray(image.data)
-        assert (len(data.shape) == len(self.shape)), \
-            'Image shape - {}, resize shape - {}'.format(data.shape, self.shape)
-
-        factor = [float(o) / i for i, o in zip(data.shape, self.shape)]
-        image.data = interpolation.zoom(data, zoom=factor, order=1)
-
-        return image
-
-
-class CropBraTS(Preprocessor):
-    __provider__ = 'crop_brats'
-
-    def process(self, image, annotation_meta=None):
-        def bbox3(img):
-            rows = np.where(np.any(np.any(img, axis=1), axis=1))
-            cols = np.where(np.any(np.any(img, axis=0), axis=1))
-            slices = np.where(np.any(np.any(img, axis=0), axis=0))
-
-            if rows[0].shape[0] > 0:
-                rmin, rmax = rows[0][[0, -1]]
-                cmin, cmax = cols[0][[0, -1]]
-                smin, smax = slices[0][[0, -1]]
-
-                return np.array([[rmin, cmin, smin], [rmax, cmax, smax]])
-            return np.array([[-1, -1, -1], [0, 0, 0]])
-
-        bboxes = np.zeros((image.data.shape[0],) + (2, 3))
-        for i in range(image.data.shape[0]):
-            bboxes[i] = bbox3(image.data[i, :, :, :])
-
-        bbox_min = np.min(bboxes[:, 0, :], axis=0).ravel().astype(int)
-        bbox_max = np.max(bboxes[:, 1, :], axis=0).ravel().astype(int)
-        bbox = np.zeros((2, 3), dtype=int)
-        bbox[0] = bbox_min
-        bbox[1] = bbox_max
-
-        image.data = image.data[:, bbox[0, 0]:bbox[1, 0], bbox[0, 1]:bbox[1, 1], bbox[0, 2]:bbox[1, 2]]
-
-        image.metadata['box'] = bbox
-
-        return image
