+++ tools/accuracy_checker/configs/brain-tumor-segmentation-0001.yml
@@ -9,6 +9,8 @@ models:
         weights: public/brain-tumor-segmentation-0001/FP32/brain-tumor-segmentation-0001.bin
         adapter:
             type: brain_tumor_segmentation
+            make_argmax: True
+            label_order: [2,1,3]
 
       - framework: dlsdk
         tags:
@@ -16,11 +18,15 @@ models:
         device: GPU
         model:   public/brain-tumor-segmentation-0001/FP16/brain-tumor-segmentation-0001.xml
         weights: public/brain-tumor-segmentation-0001/FP16/brain-tumor-segmentation-0001.bin
-        adapter: brain_tumor_segmentation
+        adapter:
+            type: brain_tumor_segmentation
+            make_argmax: True
 
     datasets:
       - name: BraTS_2017
         preprocessing:
+          - type: swap_modalities
+            modality_order: [1,3,0,2]
           - type: crop_brats
           - type: normalize_brats
             masked: True
@@ -38,7 +44,6 @@ models:
           # ground truth median             [0.9316, 0.7714, 0.8535, 0.8456]
           # UNCROPPED: ground truth median  [0.9339, 0.7918, 0.8603, 0.8576]
           - type: dice_index
-            output_order: [0,2,1,3]
             median: True
             presenter: print_vector
